WIP: split up chunk_cache_lru. start to add WASM gating
